Conditional inclusion of Rancher model spec
