+++ spec/models/rancher_spec.rb
@@ -1,3 +1,5 @@
+return unless defined?(Helm::Engine)
+
 require 'rails_helper'
 
 RSpec.describe Helm::Rancher, :type => :model do
@@ -100,65 +102,6 @@ RSpec.describe Helm::Rancher, :type => :model do
           ).and_return(true)
 
         # do it already
-        subject.save!
-      end
-    end
-
-    context 'custom install' do
-      let(:custom_config) do
-        cc = RancherOnEks::CustomConfig.new(
-          repo_name: 'mock-repo-name',
-          repo_url: 'mock-repo-url',
-          chart: 'mock-chart',
-          release_name: 'mock-release-name',
-          version: 'mock-version'
-        )
-        cc.save
-        return cc
-      end
-
-      it 'delegates with custom attributes' do
-        # create the namespace via kubectl
-        mock_kubectl = double
-        expect(mock_kubectl)
-          .to receive(:create_namespace).with(expected_namespace)
-          .and_return(true)
-        subject.instance_variable_set(:@kubectl, mock_kubectl)
-
-        # add the helm repo
-        mock_helm = double
-        expect(mock_helm)
-          .to receive(:add_repo).with(custom_config.repo_name, custom_config.repo_url)
-          .and_return(true)
-        subject.instance_variable_set(:@helm, mock_helm)
-
-        # install rancher via helm
-        expect(mock_helm)
-          .to receive(:install)
-          .with(
-            custom_config.release_name,
-            custom_config.chart,
-            expected_namespace,
-            [
-              '--set',
-              'extraEnv[0].name=CATTLE_PROMETHEUS_METRICS',
-              '--set-string',
-              'extraEnv[0].value=true',
-              '--set',
-              "hostname=#{mock_fqdn}",
-              '--set',
-              'replicas=3',
-              '--version',
-              custom_config.version
-            ]
-          ).and_return(true)
-
-        # do it already
-        subject.repo_name = custom_config.repo_name
-        subject.repo_url = custom_config.repo_url
-        subject.chart = custom_config.chart
-        subject.release_name = custom_config.release_name
-        subject.version = custom_config.version
         subject.save!
       end
     end
